requirements.txt
flask==2.3.3
gunicorn==21.2.0
tensorflow-cpu==2.13.0
pillow==10.0.0
numpy==1.24.3flask==2.3.3
gunicorn==21.2.0
tensorflow-cpu==2.13.0
pillow==10.0.0
numpy==1.24.3
